added fault based copy code. but still has a bug

+++ tests/matrix.c
@@ -3,7 +3,7 @@
 
 
 void init_matrix(matrix_t A, int value){
-	int i,j;
+	long i,j;
 	for(i=0;i<A.m;i++){
 		for(j=0;j<A.n;j++){
 			A.matrix[i*A.n+j]=value;
@@ -22,7 +22,7 @@ void increment_matrix(matrix_t A){
 }
 
 void multiply_matrix(matrix_t A, matrix_t B, matrix_t C){
-	int i,j,k;
+	long i,j,k;
 	for(i=0;i<A.m;i++){
 		for(k=0;k<B.n;k++){
 			C.matrix[i*C.n+k]=0; // init C element
@@ -36,10 +36,10 @@ void multiply_matrix(matrix_t A, matrix_t B, matrix_t C){
 }
 
 void print_matrix(matrix_t A){	
-	int i,j;
+	long i,j;
 	for(i=0;i<A.m;i++){
 		for(j=0;j<A.n;j++){
-			printf("%d  ",A.matrix[i*A.n+j]);
+			printf("%ld  ",A.matrix[i*A.n+j]);
 		}
 		printf("\n");
 	}
